moved simulate-foreground-vis to interferometer-sim to avoid cyclic dependency
